Some renaming in bip conf, remove default parameter from wif
