+++ bip_utils/conf/bip49/__init__.py
@@ -1,2 +1,2 @@
-from bip_utils.conf.bip49.bip49_coin_conf import *
+from bip_utils.conf.bip49.bip49_conf import *
 from bip_utils.conf.bip49.bip49_conf_getter import Bip49ConfGetter
