+++ bip_utils/conf/bip84/__init__.py
@@ -1,2 +1,2 @@
-from bip_utils.conf.bip84.bip84_coin_conf import *
+from bip_utils.conf.bip84.bip84_conf import *
 from bip_utils.conf.bip84.bip84_conf_getter import Bip84ConfGetter
