+++ bip_utils/conf/common/bip_conf_const.py
