+++ bip_utils/conf/common/bip_conf_enum.py
